# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Event Management API
  description: An API developed for our WBD project that serves our project.
  version: 1.0.0
servers:
  - url: http://chomspro.herokuapp.com
    description:  Main (production) Backend server
  - url: http://localhost:3000
    description:  Main (production) Frontend server
...
paths:
...
  /organizers:
    get:
      summary: Lists all the Organizers till now 
      tags: [Organizers]
      responses:
        "200":
          description: The list of all Orgnizers.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Organizers'
    post:
      summary: Adds a new Organizers record.
      tags: [Organizers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organizers'
      responses: 
        '200':
          description: Created
        '400':
          description: Error
  /organizers/{organizersID}:
    get:
      summary: Returns information of a orgnizers by ID.
      tags: [Organizers]
      parameters:
        - name: organizersID
          in: path
          required: true
          description: Autogenerated orgnizers ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a orgnizers by ID.
      tags: [Organizers]
      parameters:
        - name: organizersID
          in: path
          required: true
          description: Autogenerated orgnizers ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a orgnizers information
      tags: [Organizers]
      parameters:
        - name: organizersID
          in: path
          required: true
          description: Autogenerated orgnizers ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organizers'
      responses:
        '200':
          description: Feed Data Updated
          schema:
            $ref: "#/components/schemas/Organizers"
        '400':
          description: Error

# -------------------------------------------------------------------------
# USERS
# -------------------------------------------------------------------------
  /users:
    get:
      summary: Lists all the Users till now 
      tags: [Users]
      responses:
        "200":
          description: The list of all Users.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Users'
    post:
      summary: Adds a new User record.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses: 
        '200':
          description: Created
        '400':
          description: Error
  /users/{userID}:
    get:
      summary: Returns information of a User by ID.
      tags: [Users]
      parameters:
        - name: userID
          in: path
          required: true
          description: Autogenerated User ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a User by ID.
      tags: [Users]
      parameters:
        - name: userID
          in: path
          required: true
          description: Autogenerated User ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a User information
      tags: [Users]
      parameters:
        - name: userID
          in: path
          required: true
          description: Autogenerated user ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '200':
          description: Feed Data Updated
          schema:
            $ref: "#/components/schemas/Users"
        '400':
          description: Error

# -------------------------------------------------------------------------
# VENUE
# -------------------------------------------------------------------------
  /venue:
    get:
      summary: Lists all the Venues till now 
      tags: [Venues]
      responses:
        "200":
          description: The list of all Venues.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Venues'   
    post:
      summary: Adds a new Venue record.
      tags: [Venues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venues'
      responses: 
        '200':
          description: Created
        '400':
          description: Error
  /venue/{venueID}:
    get:
      summary: Returns information of a Venue by ID.
      tags: [Venues]
      parameters:
        - name: venueID
          in: path
          required: true
          description: Autogenerated Venue ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a Venue by ID.
      tags: [Venues]
      parameters:
        - name: venueID
          in: path
          required: true
          description: Autogenerated Venue ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a Venue information
      tags: [Venues]
      parameters:
        - name: venueID
          in: path
          required: true
          description: Autogenerated venue ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venues'
      responses:
        '200':
          description: Feed Data Updated
          schema:
            $ref: "#/components/schemas/Venues"
        '400':
          description: Error
  


























# -------------------------------------------------------------------------
# SCHEMAS
# -------------------------------------------------------------------------



#-------------------------------------------------------------------
...